Redirect to Replicate instead of iframe
I have also updated some of the descriptions / readmes on Replicate to
reflect the content that was on these pages.

The only one that's a bit weird is the genesis model, which doesn't
actually do what it says it does any longer, so I have just left that
one blank.